refactor: minor improvements in shared enums; docs
