+++ docs/features/tasks.md
@@ -7,7 +7,7 @@ AlgoKit Tasks are a collection of handy tasks that can be used to perform variou
 - [Wallet Aliasing](./tasks/wallet.md) - Manage your Algorand addresses and accounts effortlessly with the AlgoKit Wallet feature. This feature allows you to create short aliases for your addresses and accounts on AlgoKit CLI.
 - [Vanity Address Generation](./tasks/vanity.md) - Generate vanity addresses for your Algorand accounts with the AlgoKit Vanity feature. This feature allows you to generate Algorand addresses with a custom prefix of your choice.
 - [Transfer Assets or Algos](./tasks/transfer.md) - Transfer Algos or Assets from one account to another with the AlgoKit Transfer feature. This feature allows you to transfer Algos or Assets from one account to another on Algorand blockchain.
-- [Opt-(in|out) Assets](./task/opt.md) - Opt-in or opt-out of Algorand Asset(s). Supports single or multiple assets.
+- [Opt-(in|out) Assets](./tasks/opt.md) - Opt-in or opt-out of Algorand Asset(s). Supports single or multiple assets.
 - [Signing transactions](./tasks/sign.md) - Sign goal clerk compatible Algorand transactions.
 - [Sending transactions](./tasks/send.md) - Send signed goal clerk compatible Algorand transactions.
 - [NFD lookups](./tasks/nfd.md) - Perform a lookup via NFD domain or address, returning the associated address or domain respectively using the AlgoKit CLI.
