+++ docs/features/tasks/nfd.md
@@ -39,3 +39,7 @@ $ algokit task nfd-lookup {NFD_DOMAIN_OR_ALGORAND_ADDRESS} --output full
 ```
 
 If the lookup is successful, the result will be output to the console in a JSON format.
+
+## Further Reading
+
+For in-depth details, visit the [nfd-lookup section](../../cli/index.md#nfd-lookup) in the AlgoKit CLI reference documentation.
