+++ docs/features/tasks/vanity_address.md
@@ -50,3 +50,7 @@ Generate a vanity address with the keyword "ALGO" at the start of the address an
 ```bash
 $ ~ algokit task vanity-address ALGO -o alias -a my-vanity-address
 ```
+
+## Further Reading
+
+For in-depth details, visit the [vanity-address section](../../cli/index.md#vanity-address) in the AlgoKit CLI reference documentation.
